(#85) Organizado métodos e files necessarios
