+++ src/index.js
@@ -1,5 +1,6 @@
 import {criarPullRequest} from './services.js';
-
+import {CodigoDevicePost} from './services.js';
+import {ConfirmaLoginContaUsuario} from './services.js';
 //URL DAS PÁGINAS
 var currentUrl = String(window.location.href);
 var urlParts = currentUrl.split("/");
@@ -10,12 +11,10 @@ if(localStorage.getItem('abrirModal')){
   localStorage.removeItem('abrirModal');
 }
 
-
 document.addEventListener("click", function (e) {
-  
+
     //VALIDA ENTRADA EM GITHUB MILESTONES
     if (urlParts[5] == "milestones" && urlParts[2] == "github.com") {
-  
     //BOTOES    
     var closeMilestoneButton = document.querySelector(".d-inline-block.mr-2 .btn-link");
     var getFormAction = document.querySelectorAll('form[class="d-inline-block mr-2"]')[0].action;
@@ -24,7 +23,6 @@ document.addEventListener("click", function (e) {
     //VALIDA FECHAMENTO DE MILESTONE
     if (e.path[0] == closeMilestoneButton &&
         closeMilestoneButton.textContent.toLowerCase() == "close") {
-        
         //INSTRUÇÃO PARA AUTORIZAR A ABERTURA DO MODAL
         localStorage.setItem('abrirModal', 'true')
         var base = document.querySelectorAll("form.d-inline-block.mr-2").action;
@@ -34,64 +32,16 @@ document.addEventListener("click", function (e) {
   });
 
  
-
-
   if(urlParts[2] == "github.com" && urlParts[3] == "login" && urlParts[4] == "device"  && urlParts[5] != "success" &&  urlParts[5] != "confirmation"){
     
-    var xhr = new XMLHttpRequest();
-
-    xhr.addEventListener("readystatechange", function() {
-      if(this.readyState === 4) {
-
-         
-        
-        const $getResponse = String(this.responseText).split("&");
-        const $separeteValueDevice = $getResponse[0].split("=");
-        var deviceCode = $separeteValueDevice[1];
-        localStorage.setItem("key_dev", deviceCode);
-       const $separatevalue = String($getResponse[3]).split("=");
-       var codeUser = $separatevalue[1];
-        loginDevice(codeUser)
-      }
-    });
-
-    xhr.open("POST", "https://github.com/login/device/code?client_id=46da77694ca94b6a86d7&scope=repo%20user");
-    xhr.send();
+    CodigoDevicePost();
   }
 
   if(urlParts[2] == "github.com" && urlParts[3] == "login" && urlParts[4] == "device" && urlParts[5] == "success" ){
-
-    var keyValue = localStorage.getItem('key_dev');
-    var xhr = new XMLHttpRequest();
-
-    xhr.addEventListener("readystatechange", function() {
-      if(this.readyState === 4) {
-        const $getResponse = String(this.responseText).split("&");
-        const $separateToken = $getResponse[0].split("=");
-        const $token = $separateToken[1]
-        localStorage.setItem("token",$token);
-      }
-    });
-
-    xhr.open("POST", "https://github.com/login/oauth/access_token?client_id=46da77694ca94b6a86d7&device_code="+keyValue+"&grant_type=urn:ietf:params:oauth:grant-type:device_code");
-    xhr.setRequestHeader("Cookie", "_device_id="+keyValue+"; _gh_sess=eJ4SGU%2B24YwcePUSGoD82e%2Bm5C6pCMXSTicqWhqxjqK2q%2FdT%2BnIZRy2VT17XgsvqnrMmPA8Nm3foDTdRakuDTl5ZVWW%2F%2BzDAqvsEv6B4eoE83TNIIaC1HTZvT93ojGs8ARDPh26hSSFoUIYDHYnz6KEDJTK3hjlrX5ATkhOLy%2BDUwus78XVkwT1hhC0nIpj4Qm0YC6Z2rqNGkDeCzg6Kz5O0tSNzQxyahCC9WHLgoECgonCtnYIQ27w%2BBmzQ%2FWIqRmaUM%2BPRUOZBGmttLbSEhX4FaHUNIlglvwjNx9R34UAcaVkO--YFp0BPOo4pjLQIHV--aQrfy79bV3QfnmGMwYPhGw%3D%3D; _octo=GH1.1.2068834575.1616879702; logged_in=no");
-
-    xhr.send();
-    
+   
+    ConfirmaLoginContaUsuario();
     
   }
   
   
-  function loginDevice(codeUser){
-   
-    document.getElementById("user-code-0").value = codeUser[0];
-    document.getElementById("user-code-1").value = codeUser[1];
-    document.getElementById("user-code-2").value =  codeUser[2];
-    document.getElementById("user-code-3").value = codeUser[3];
-    document.getElementById("user-code-5").value = codeUser[5];
-    document.getElementById("user-code-6").value = codeUser[6];
-    document.getElementById("user-code-7").value = codeUser[7];
-    document.getElementById("user-code-8").value = codeUser[8];
-    
-    
-}
+
