+++ src/popup.html
@@ -6,6 +6,6 @@
   <body>
     <h1>Git bot Extension</h1>
     <button id="btnLink">Connect</button>
-    <script src="OauthLogIn.js"></script>
+    <script src="popup.js"></script>
   </body>
 </html>
